Added ageExact function
Added sample-level, series-level and text-level age extraction functionality.
Tried to update all the docstrings.

Shifted some processing from mid-run to the end of the scraper. I'd like to start pushing toward more of this architecture because scraping and processing simultaneously is unnecessarily complicated and causes failures resulting in time and data loss.